Renamed visualizations and changed column size
