--- plugins/main/public/components/common/wazuh-discover/config/data-grid-columns.tsx
@@ -1,11 +1,0 @@
-import { tDataGridColumn } from '../../data-grid';
-import { commonColumns } from '../../../overview/common/data-grid-columns';
-
-export const threatHuntingColumns: tDataGridColumn[] = [
-  commonColumns.timestamp,
-  commonColumns['wazuh.agent.name'],
-  { id: 'wazuh.integration.category', initialWidth: 130 },
-  { id: 'wazuh.integration.name' },
-  { id: 'wazuh.rule.title' },
-  { id: 'wazuh.rule.level', initialWidth: 150 },
-];
