+++ plugins/main/public/components/overview/threat-hunting/events/threat-hunting-columns.tsx
@@ -1,39 +1,11 @@
-import { EuiDataGridColumn } from '@elastic/eui';
 import { tDataGridColumn } from '../../../common/data-grid';
 import { commonColumns } from '../../common/data-grid-columns';
 
-export const MAX_ENTRIES_PER_QUERY = 10000;
-
-const threatHuntingTableCommonColumns = {
-  icon: {
-    id: 'icon',
-  },
-  'wazuh.rule.mitre.id': {
-    id: 'wazuh.rule.mitre.id',
-    initialWidth: 123,
-  },
-  'wazuh.rule.mitre.tactic': {
-    id: 'wazuh.rule.mitre.tactic',
-    initialWidth: 266,
-  },
-} as Record<string, tDataGridColumn>;
-
-export const threatHuntingTableDefaultColumns: tDataGridColumn[] = [
-  threatHuntingTableCommonColumns.icon,
-  commonColumns['wazuh.agent.id'],
-  threatHuntingTableCommonColumns['wazuh.rule.mitre.id'],
-  threatHuntingTableCommonColumns['wazuh.rule.mitre.tactic'],
-  commonColumns['wazuh.rule.description'],
-  commonColumns['wazuh.rule.level'],
-  commonColumns['wazuh.rule.id'],
-];
-
-export const threatHuntingTableAgentColumns: EuiDataGridColumn[] = [
-  threatHuntingTableCommonColumns.icon,
+export const threatHuntingColumns: tDataGridColumn[] = [
   commonColumns.timestamp,
-  threatHuntingTableCommonColumns['wazuh.rule.mitre.id'],
-  threatHuntingTableCommonColumns['wazuh.rule.mitre.tactic'],
-  commonColumns['wazuh.rule.description'],
-  commonColumns['wazuh.rule.level'],
-  commonColumns['wazuh.rule.id'],
+  commonColumns['wazuh.agent.name'],
+  { id: 'wazuh.integration.category', initialWidth: 230 },
+  { id: 'wazuh.integration.name' },
+  { id: 'wazuh.rule.title' },
+  { id: 'wazuh.rule.level', initialWidth: 150 },
 ];
